python: refactor output formatting code
Problem: The job-specific JobInfoFormat class embeds JobFormatter
and HeaderFormat classes and a filter_empty() method, all of which
are not necessarily job-specific classes and methods. This means that
other utilities and code that generate formatted output can't reuse
these classes and methods.

Move the JobFormatter class to flux.util.UtilFormatter and embed the
generic HeaderFormatter class in flux.util.OutputFormat. Also move
the useful filter_empty() method to flux.util.OutputFormat where it
can be reused.

Also move JobDatetime, which is also not really job-specific, to
flux.util.UtilDatetime, and the fsd() function from flux.job.info
into flux.util.
