+++ src/bindings/python/flux/job/info.py
@@ -13,7 +13,6 @@ import time
 import pwd
 import json
 import string
-from datetime import datetime, timedelta
 from collections import namedtuple
 
 import flux.constants
@@ -371,7 +370,7 @@ class JobInfo:
                 if eta < 0:
                     eta = "now"
                 else:
-                    eta = fsd(eta)
+                    eta = flux.util.fsd(eta)
                 return f"eta:{eta}"
             except TypeError:
                 return ""
@@ -462,144 +461,12 @@ def job_fields_to_attrs(fields):
     return attrs
 
 
-def fsd(secs):
-    #  Round <1ms down to 0s for now
-    if secs < 1.0e-3:
-        strtmp = "0s"
-    elif secs < 10.0:
-        strtmp = "%.03fs" % secs
-    elif secs < 60.0:
-        strtmp = "%.4gs" % secs
-    elif secs < (60.0 * 60.0):
-        strtmp = "%.4gm" % (secs / 60.0)
-    elif secs < (60.0 * 60.0 * 24.0):
-        strtmp = "%.4gh" % (secs / (60.0 * 60.0))
-    else:
-        strtmp = "%.4gd" % (secs / (60.0 * 60.0 * 24.0))
-    return strtmp
-
-
-class JobDatetime(datetime):
-    """
-    Subclass of datetime but with a __format__ method that supports
-    width and alignment specification after any time formatting via
-    two colons `::` before the Python format spec, e.g::
-
-    >>> "{0:%b%d %R::>16}".format(JobDatetime.now())
-    '     Sep21 18:36'
-    """
-
-    def __format__(self, fmt):
-        # The string "::" is used to split the strftime() fromat from
-        # any Python format spec:
-        vals = fmt.split("::", 1)
-
-        # Call strftime() to get the formatted datetime as a string
-        result = self.strftime(vals[0])
-
-        # If there was a format spec, apply it here:
-        try:
-            return f"{{0:{vals[1]}}}".format(result)
-        except IndexError:
-            return result
-
-
 class JobInfoFormat(flux.util.OutputFormat):
     """
     Store a parsed version of an output format string for JobInfo objects,
     allowing the fields to iterated without modifiers, building
     a new format suitable for headers display, etc...
     """
-
-    class JobFormatter(string.Formatter):
-        # pylint: disable=too-many-branches
-        def convert_field(self, value, conv):
-            """
-            Flux job-specific field conversions. Avoids the need
-            to create many different format field names to represent
-            different conversion types. (mainly used for time-specific
-            fields for now).
-            """
-            orig_value = str(value)
-            if conv == "d":
-                # convert from float seconds since epoch to a datetime.
-                # User can than use datetime specific format fields, e.g.
-                # {t_inactive!d:%H:%M:%S}.
-                try:
-                    value = JobDatetime.fromtimestamp(value)
-                except TypeError:
-                    if orig_value == "":
-                        value = JobDatetime.fromtimestamp(0.0)
-                    else:
-                        raise
-            elif conv == "D":
-                # As above, but convert to ISO 8601 date time string.
-                try:
-                    value = datetime.fromtimestamp(value).strftime("%FT%T")
-                except TypeError:
-                    if orig_value == "":
-                        value = ""
-                    else:
-                        raise
-            elif conv == "F":
-                # convert to Flux Standard Duration (fsd) string.
-                try:
-                    value = fsd(value)
-                except TypeError:
-                    if orig_value == "":
-                        value = ""
-                    else:
-                        raise
-            elif conv == "H":
-                # if > 0, always round up to at least one second to
-                #  avoid presenting a nonzero timedelta as zero
-                try:
-                    if 0 < value < 1:
-                        value = 1
-                    value = str(timedelta(seconds=round(value)))
-                except TypeError:
-                    if orig_value == "":
-                        value = ""
-                    else:
-                        raise
-            elif conv == "P":
-                #  Convert a floating point to percentage
-                try:
-                    value = value * 100
-                    if 0 < value < 1:
-                        value = f"{value:.2f}%"
-                    else:
-                        value = f"{value:.3g}%"
-                except (TypeError, ValueError):
-                    if orig_value == "":
-                        value = ""
-                    else:
-                        raise
-            else:
-                value = super().convert_field(value, conv)
-            return value
-
-        def format_field(self, value, spec):
-            if spec.endswith("h"):
-                basecases = ("", "0s", "0.0", "0:00:00", "1970-01-01T00:00:00")
-                value = "-" if str(value) in basecases else str(value)
-                spec = spec[:-1] + "s"
-            return super().format_field(value, spec)
-
-    class HeaderFormatter(JobFormatter):
-        """Custom formatter for flux-jobs(1) header row.
-
-        Override default formatter behavior of calling getattr() on dotted
-        field names. Instead look up header literally in kwargs.
-        This greatly simplifies header name registration as well as
-        registration of "valid" fields.
-        """
-
-        def get_field(self, field_name, args, kwargs):
-            """Override get_field() so we don't do the normal gettatr thing"""
-            if field_name in kwargs:
-                return kwargs[field_name], None
-            return super().get_field(field_name, args, kwargs)
 
     #  List of legal format fields and their header names
     headings = {
@@ -707,70 +574,3 @@ class JobInfoFormat(flux.util.OutputFormat):
                         field_heading = field_heading[10:]
                     self.headings[field] = field_heading
         super().__init__(self.headings, fmt, prepend="0.")
-
-    def format(self, obj):
-        """
-        format object with our JobFormatter
-        """
-        return self.JobFormatter().format(self.get_format(), obj)
-
-    def header(self):
-        """
-        format header with custom HeaderFormatter
-        """
-        return self.HeaderFormatter().format(self.header_format(), **self.headings)
-
-    def filter_empty(self, items):
-        """
-        Check for format fields that are prefixed with `?:` (e.g. "?:{name}")
-        and filter them out of the current format string if they result in an
-        empty string for every entry in `items`.
-        """
-        #  Build a list of all format strings that have the collapsible
-        #  sentinel '?:' to determine which are subject to the test for
-        #  emptiness.
-        #
-        #  Note: we remove the leading `text` and the format `spec` because
-        #  these may make even empty formatted fields non-empty. E.g. a spec
-        #  of `:>8` will always make the format result 8 characters wide.
-        #
-        lst = []
-        index = 0
-        for (text, field, _, conv) in self.format_list:
-            if text.endswith("?:"):
-                fmt = self._fmt_tuple("", "0." + field, None, conv)
-                lst.append(dict(fmt=fmt, index=index))
-            index = index + 1
-
-        # Return immediately if no format fields are collapsible
-        if not lst:
-            return self
-
-        formatter = self.JobFormatter()
-
-        #  Iterate over all items, rebuilding lst each time to contain
-        #  only those fields that resulted in nonzero strings:
-        for item in items:
-            lst = [x for x in lst if not formatter.format(x["fmt"], item)]
-
-            #  If lst is now empty, that means all fields already returned
-            #  nonzero strings, so we can break early
-            if not lst:
-                break
-
-        #  Remove any entries that were empty from self.format_list
-        #  (use index field of lst to remove by postition in self.format_list)
-        format_list = [
-            x
-            for i, x in enumerate(self.format_list)
-            if i not in [x["index"] for x in lst]
-        ]
-
-        #  Remove "?:" from remaining entries so they disappear in ouput.
-        for entry in format_list:
-            if entry[0].endswith("?:"):
-                entry[0] = entry[0][:-2]
-
-        #  Return new format string created from pruned format_list
-        fmt = "".join(self._fmt_tuple(*x) for x in format_list)
-        return JobInfoFormat(fmt)
